requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
Flask-Login==0.6.2
requests==2.28.1
python-dotenv==0.20.0